feat: update request validation

+++ routes/api.php
@@ -33,8 +33,9 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::group(['prefix' => 'users'], function () {
         Route::get('/', [UserController::class, 'index']);
         Route::get('/{user}', [UserController::class, 'show']);
-        Route::post('/', [UserController::class, 'store']);
-        Route::put('/{user}', [UserController::class, 'update']);
-        Route::delete('/{user}', [UserController::class, 'destroy']);
+
+        Route::post('/', [UserController::class, 'store'])->middleware([CheckUserRole::class . ':Admin,Manager']);
+        Route::put('/{user}', [UserController::class, 'update'])->middleware([CheckUserRole::class . ':Admin,Manager']);
+        Route::delete('/{user}', [UserController::class, 'destroy'])->middleware([CheckUserRole::class . ':Admin,Manager']);
     });
 });
